Refactor project architecture to allow multiple jars in /lib
