+++ graphalytics-platforms-openg-granula/src/main/java/nl/tudelft/graphalytics/openg/reporting/logging/OpenGLogger.java
