+++ graphalytics-platforms-openg-granula/src/main/resources/openg.platform
@@ -1,0 +1,1 @@
+nl.tudelft.graphalytics.openg.OpenGGranulaPlatform
