+++ graphalytics-platforms-openg-std/src/main/c/CMakeLists.txt
