+++ graphalytics-platforms-openg-std/src/main/c/bfs.cpp
