+++ graphalytics-platforms-openg-std/src/main/c/lcc.cpp
