+++ graphalytics-platforms-openg-std/src/main/c/pr.cpp
