+++ graphalytics-platforms-openg-std/src/main/c/wcc.cpp
