+++ graphalytics-platforms-openg-std/src/main/java/nl/tudelft/graphalytics/openg/OpenGJob.java
