+++ graphalytics-platforms-openg-std/src/main/java/nl/tudelft/graphalytics/openg/algorithms/cdlp/CommunityDetectionLPJob.java
