+++ graphalytics-platforms-openg-std/src/main/java/nl/tudelft/graphalytics/openg/algorithms/pr/PageRankJob.java
