+++ graphalytics-platforms-openg-std/src/main/java/nl/tudelft/graphalytics/openg/config/JobConfiguration.java
