+++ graphalytics-platforms-openg-std/src/main/resources/log4j2.xml
