+++ graphalytics-platforms-openg-std/src/main/resources/openg.platform
