+++ pom.xml
@@ -2,30 +2,18 @@
 	<modelVersion>4.0.0</modelVersion>
 
 	<groupId>nl.tudelft.graphalytics</groupId>
-	<artifactId>graphalytics-platforms-openg</artifactId>
+	<artifactId>graphalytics-platforms-openg-root</artifactId>
 	<version>0.1-SNAPSHOT</version>
-	<packaging>jar</packaging>
+	<packaging>pom</packaging>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-
-		<commons.exec.version>1.3</commons.exec.version>
-		<fastutil.version>7.0.7</fastutil.version>
 		<graphalytics.version>0.3-SNAPSHOT</graphalytics.version>
-		<hamcrest.version>1.3</hamcrest.version>
-		<junit.version>4.11</junit.version>
-		<log4j.version>2.5</log4j.version>
 	</properties>
 
-	<!--
-	<scm>
-		<developerConnection>scm:git:ssh://git@github.com/tudelft-atlarge/graphalytics-platforms-openg</developerConnection>
-		<tag>HEAD</tag>
-	</scm>
-	-->
 
 	<dependencies>
-               <!-- Graphalytics dependencies: core + resources, validation -->
+		<!-- Graphalytics dependencies: core + resources, validation -->
 		<dependency>
 			<groupId>nl.tudelft.graphalytics</groupId>
 			<artifactId>graphalytics-core</artifactId>
@@ -45,135 +33,27 @@
 			<version>${graphalytics.version}</version>
 			<scope>test</scope>
 		</dependency>
-
-		<!-- Granula dependencies -->
-		<dependency>
-			<groupId>nl.tudelft.graphalytics</groupId>
-			<artifactId>graphalytics-plugins-granula</artifactId>
-			<version>${graphalytics.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>nl.tudelft.pds.granula</groupId>
-			<artifactId>openg-analyzer</artifactId>
-			<version>0.1.2</version>
-		</dependency>
-
-		<dependency>
-			<groupId>org.apache.logging.log4j</groupId>
-			<artifactId>log4j-core</artifactId>
-			<version>${log4j.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.logging.log4j</groupId>
-			<artifactId>log4j-slf4j-impl</artifactId>
-			<version>${log4j.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.apache.commons</groupId>
-			<artifactId>commons-exec</artifactId>
-			<version>${commons.exec.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>it.unimi.dsi</groupId>
-			<artifactId>fastutil</artifactId>
-			<version>${fastutil.version}</version>
-		</dependency>
-
-		<dependency>
-			<groupId>junit</groupId>
-			<artifactId>junit</artifactId>
-			<version>${junit.version}</version>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.hamcrest</groupId>
-			<artifactId>hamcrest-all</artifactId>
-			<version>${hamcrest.version}</version>
-			<scope>test</scope>
-		</dependency>
 	</dependencies>
+
+
+	<modules>
+		<module>graphalytics-platforms-openg-std</module>
+	</modules>
+
+	<profiles>
+		<profile>
+			<id>granula</id>
+			<activation>
+				<activeByDefault>false</activeByDefault>
+			</activation>
+			<modules>
+				<module>graphalytics-platforms-openg-granula</module>
+			</modules>
+		</profile>
+	</profiles>
 
 	<build>
 		<plugins>
-			<!-- Java compiler settings -->
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-compiler-plugin</artifactId>
-				<version>3.1</version>
-				<configuration>
-					<source>1.7</source>
-					<target>1.7</target>
-				</configuration>
-			</plugin>
-
-			<!-- Maven Shade plugin used by platform modules to create fat JARs -->
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-shade-plugin</artifactId>
-				<version>2.3</version>
-				<configuration>
-					<createDependencyReducedPom>false</createDependencyReducedPom>
-					<minimizeJar>false</minimizeJar>
-					<artifactSet>
-						<excludes>
-							<exclude>*:*:*:resources</exclude>
-						</excludes>
-					</artifactSet>
-					<filters>
-						<filter>
-							<artifact>*.*</artifact>
-							<excludes>
-								<exclude>META-INF/*.SF</exclude>
-								<exclude>META-INF/*.DSA</exclude>
-								<exclude>META-INF/*.RSA</exclude>
-							</excludes>
-						</filter>
-					</filters>
-					<finalName>${artifactId}-${version}-original-final</finalName>
-				</configuration>
-				<executions>
-					<execution>
-						<phase>package</phase>
-						<goals>
-							<goal>shade</goal>
-						</goals>
-					</execution>
-				</executions>
-			</plugin>
-
-			<!-- License plugin for checking that all source files have the appropriate license -->
-			<plugin>
-				<groupId>com.mycila</groupId>
-				<artifactId>license-maven-plugin</artifactId>
-				<version>2.10</version>
-				<configuration>
-					<header>http://www.pds.ewi.tudelft.nl/~mihai/copyright-notice-template</header>
-					<properties>
-						<owner>Delft University of Technology</owner>
-						<project.inceptionYear>2015</project.inceptionYear>
-					</properties>
-					<excludes>
-						<exclude>**/README</exclude>
-						<exclude>**/LICENSE</exclude>
-						<exclude>**/pom.xml</exclude>
-						<exclude>src/main/assembly/**</exclude>
-						<exclude>**/build.sh</exclude>
-						<exclude>src/main/resources/**</exclude>
-						<exclude>config-template/**</exclude>
-					</excludes>
-					<mapping>
-						<java>SLASHSTAR_STYLE</java>
-					</mapping>
-				</configuration>
-				<executions>
-					<execution>
-						<goals>
-							<goal>check</goal>
-						</goals>
-					</execution>
-				</executions>
-			</plugin>
-
 			<!-- Plugin for creating distribution archives -->
 			<plugin>
 				<groupId>org.apache.maven.plugins</groupId>
@@ -188,6 +68,7 @@
 						</goals>
 						<configuration>
 							<descriptors>
+								<!--<descriptor>src/main/assembly/bin.xml</descriptor>-->
 								<descriptor>src/main/assembly/bin.xml</descriptor>
 							</descriptors>
 							<finalName>graphalytics-${graphalytics.version}-openg-${project.version}</finalName>
@@ -197,21 +78,8 @@
 					</execution>
 				</executions>
 			</plugin>
-
-			<!-- Maven release management -->
-			<!--
-			<plugin>
-				<groupId>org.apache.maven.plugins</groupId>
-				<artifactId>maven-release-plugin</artifactId>
-				<version>2.5.1</version>
-				<configuration>
-					<tagNameFormat>v@{project.version}</tagNameFormat>
-					<autoVersionSubmodules>true</autoVersionSubmodules>
-					<pushChanges>false</pushChanges>
-				</configuration>
-			</plugin>
-			-->
 		</plugins>
 	</build>
+
 </project>
