+++ src/main/assembly/bin.xml
@@ -1,56 +1,76 @@
 <assembly xmlns="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2"
-		xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2 http://maven.apache.org/xsd/assembly-1.1.2.xsd">
-	<id>bin</id>
-	<formats>
-		<format>tar.gz</format>
-	</formats>
+          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+          xsi:schemaLocation="http://maven.apache.org/plugins/maven-assembly-plugin/assembly/1.1.2 http://maven.apache.org/xsd/assembly-1.1.2.xsd">
+    <id>bin</id>
+    <formats>
+        <format>tar.gz</format>
+    </formats>
 
-	<fileSets>
-		<fileSet>
-			<directory>${project.basedir}</directory>
-			<outputDirectory>.</outputDirectory>
-			<includes>
-				<include>prepare-benchmark.sh</include>
-				<include>config-template/**</include>
-			</includes>
-		</fileSet>
-		<fileSet>
-			<directory>${project.build.directory}</directory>
-			<outputDirectory>lib</outputDirectory>
-			<includes>
-				<include>*final.jar</include>
-			</includes>
-			<excludes>
-				<exclude>original*</exclude>
-			</excludes>
-		</fileSet>
-		<fileSet>
-			<directory>src/main/c</directory>
-			<outputDirectory>src</outputDirectory>
-		</fileSet>
-	</fileSets>
-	<files>
-		<file>
-			<source>${project.basedir}/README.md</source>
-			<outputDirectory>.</outputDirectory>
-			<destName>README-graphalytics-openg.md</destName>
-		</file>
-	</files>
+    <files>
+        <!--Platform ReadMe files-->
+        <file>
+            <source>${project.basedir}/README.md</source>
+            <outputDirectory>.</outputDirectory>
+            <destName>README-graphalytics-openg.md</destName>
+        </file>
+    </files>
+    <fileSets>
+        <!--Platform configuration files-->
+        <fileSet>
+            <directory>${project.basedir}</directory>
+            <outputDirectory>.</outputDirectory>
+            <includes>
+                <include>prepare-benchmark.sh</include>
+                <include>config-template/**</include>
+            </includes>
+        </fileSet>
+        <!--Native C code needed to be compiled at runtime-->
+        <fileSet>
+            <directory>${basedir}/graphalytics-platforms-openg-std/src</directory>
+            <outputDirectory>src</outputDirectory>
+        </fileSet>
+        <fileSet>
+            <directory>${basedir}/graphalytics-platforms-openg-granula/src</directory>
+            <outputDirectory>src</outputDirectory>
+        </fileSet>
+        <!--Graphalytics library jars-->
+        <fileSet>
+            <directory>${basedir}/graphalytics-platforms-openg-std/target/</directory>
+            <outputDirectory>lib</outputDirectory>
+            <includes>
+                <include>graphalytics-platform*.jar</include>
+            </includes>
+            <excludes>
+                <exclude>original*.jar</exclude>
+            </excludes>
+        </fileSet>
+        <fileSet>
+            <directory>${basedir}/graphalytics-platforms-openg-granula/target/</directory>
+            <outputDirectory>lib</outputDirectory>
+            <includes>
+                <include>graphalytics-platform*.jar</include>
+            </includes>
+            <excludes>
+                <exclude>original*.jar</exclude>
+            </excludes>
+        </fileSet>
+    </fileSets>
 
-	<dependencySets>
-		<dependencySet>
-			<outputDirectory>.</outputDirectory>
-			<includes>
-				<include>*:resources</include>
-			</includes>
-			<unpack>true</unpack>
-			<unpackOptions>
-				<excludes>
-					<exclude>META-INF/**</exclude>
-				</excludes>
-			</unpackOptions>
-		</dependencySet>
-	</dependencySets>
+
+    <!--Core common files-->
+    <dependencySets>
+        <dependencySet>
+            <outputDirectory>.</outputDirectory>
+            <includes>
+                <include>*:resources</include>
+            </includes>
+            <unpack>true</unpack>
+            <unpackOptions>
+                <excludes>
+                    <exclude>META-INF/**</exclude>
+                </excludes>
+            </unpackOptions>
+        </dependencySet>
+    </dependencySets>
 
 </assembly>
